+++ src/app/widgets/SOU/sou-streams/config.ts
@@ -1,0 +1,38 @@
+export const TITLES_OF_TABLE: { name: string, widthOfBlock: string }[] = [
+    {
+        name: 'Начало операции',
+        widthOfBlock: '138px'
+    },
+    {
+        name: 'Конец операции',
+        widthOfBlock: '138px'
+    },
+    {
+        name: 'Тип источника',
+        widthOfBlock: '138px'
+    },
+    {
+        name: 'Источник',
+        widthOfBlock: '138px'
+    },
+    {
+        name: 'Поток',
+        widthOfBlock: '138px'
+    },
+    {
+        name: 'Продукт в источнике',
+        widthOfBlock: '196px'
+    },
+    {
+        name: 'Тип приёмника',
+        widthOfBlock: '138px'
+    },
+    {
+        name: 'Приёмник',
+        widthOfBlock: '138px'
+    },
+    {
+        name: 'Продукт в приёмнике',
+        widthOfBlock: '196px'
+    }
+];