+++ src/app/widgets/admin/admin-report-server-configurator/components/admin-report-server-configurator-repository/admin-report-server-configurator-repository-address/admin-report-server-configurator-repository-address.component.ts
@@ -1,4 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
+import { IChildrenFolder } from '@dashboard/models/ADMIN/report-server.model';
+import { ITemplateFolder } from '@widgets/admin/admin-report-server-configurator/models/admin-report-server-configurator.model';
+import { AdminReportServerConfiguratorRootService } from '@widgets/admin/admin-report-server-configurator/services/admin-report-server-configurator-root.service';
+import { AdminReportConfiguratorService } from '@widgets/admin/admin-report-server-configurator/services/admin-report-server-configurator.service';
 
 @Component({
   selector: 'evj-admin-report-server-configurator-repository-address',
@@ -7,9 +11,35 @@ import { Component, OnInit } from '@angular/core';
 })
 export class AdminReportServerConfiguratorRepositoryAddressComponent implements OnInit {
 
-  constructor() { }
+  public readonly arrowsIcon="assets/icons/widgets/admin/admin-report-server-configurator/path-arrows.svg";
+  public path: ITemplateFolder[] = [];
+
+  constructor(
+    public arscService: AdminReportConfiguratorService,
+    public arscRootService: AdminReportServerConfiguratorRootService
+  ) { }
 
   ngOnInit(): void {
+    this.arscService.address$.subscribe(value => {
+      if (value) {
+        this.path.push(value);
+      }      
+    })
+    console.log(this.path);
+    
   }
 
+  public openFolder(item: ITemplateFolder ): void {
+    this.arscService.folders$.next(item.childFolders);
+    this.arscService.reports$.next(item.templates);
+    const idx = this.path.findIndex(v => v?.id === item?.id);
+    this.path.splice(idx + 1, this.path.length);
+  }
+
+  public async mainPage(): Promise<void> {
+    this.path = [];
+    const data = await this.arscRootService.getTemplateFolder();
+    this.arscService.folders$.next(data.folders);
+    this.arscService.reports$.next(data.templates);
+  }
 }